+++ back/mall-common/src/main/java/com/syh/common/product/dtos/ProductListDto.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class ProductListDto {
     Integer current;
     Integer size;
-    Integer categoryId;
+    Long categoryId;
 }
