+++ back/mall-services/mall-customer-service/src/main/java/com/syh/service/customer/services/CustomerService.java
@@ -11,5 +11,5 @@ import com.syh.common.customer.pojos.CustomerInf;
 */
 public interface CustomerService extends IService<CustomerInf> {
 
-    Result<CustomerInf> customerOverview(Integer customerId);
+    Result<CustomerInf> customerOverview(Long customerId);
 }
