+++ back/mall-services/mall-product-service/src/main/java/com/syh/service/product/feign/OrderFeignClient.java
@@ -9,5 +9,5 @@ import org.springframework.web.bind.annotation.RequestParam;
 @FeignClient(name = "mall-customer-service")
 public interface OrderFeignClient {
     @RequestMapping("/customerOverview")
-    Result<CustomerInf> customerOverview(@RequestParam("customerId") Integer customerId);
+    Result<CustomerInf> customerOverview(@RequestParam("customerId") Long customerId);
 }
